+++ app/src/main/java/eu/faircode/email/FragmentAnswers.java
@@ -57,6 +57,7 @@ import java.util.Objects;
 public class FragmentAnswers extends FragmentBase {
     private boolean cards;
 
+    private View view;
     private RecyclerView rvAnswer;
     private ContentLoadingProgressBar pbWait;
     private Group grpReady;
@@ -79,7 +80,7 @@ public class FragmentAnswers extends FragmentBase {
         setSubtitle(R.string.menu_answers);
         setHasOptionsMenu(true);
 
-        View view = inflater.inflate(R.layout.fragment_answers, container, false);
+        view = inflater.inflate(R.layout.fragment_answers, container, false);
 
         // Get controls
         rvAnswer = view.findViewById(R.id.rvAnswer);
@@ -228,10 +229,21 @@ public class FragmentAnswers extends FragmentBase {
         SearchView searchView = (SearchView) menuSearch.getActionView();
         searchView.setQueryHint(getString(R.string.title_rules_search_hint));
 
-        if (!TextUtils.isEmpty(searching)) {
-            menuSearch.expandActionView();
-            searchView.setQuery(searching, true);
-        }
+        final String search = searching;
+        view.post(new RunnableEx("answers:search") {
+            @Override
+            public void delegate() {
+                if (!getLifecycle().getCurrentState().isAtLeast(Lifecycle.State.STARTED))
+                    return;
+
+                if (TextUtils.isEmpty(search))
+                    menuSearch.collapseActionView();
+                else {
+                    menuSearch.expandActionView();
+                    searchView.setQuery(search, true);
+                }
+            }
+        });
 
         getViewLifecycleOwner().getLifecycle().addObserver(new LifecycleObserver() {
             @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
@@ -244,7 +256,7 @@ public class FragmentAnswers extends FragmentBase {
         searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
             @Override
             public boolean onQueryTextChange(String newText) {
-                if (getView() != null) {
+                if (getLifecycle().getCurrentState().isAtLeast(Lifecycle.State.STARTED)) {
                     searching = newText;
                     adapter.search(newText);
                 }
@@ -253,8 +265,10 @@ public class FragmentAnswers extends FragmentBase {
 
             @Override
             public boolean onQueryTextSubmit(String query) {
-                searching = query;
-                adapter.search(query);
+                if (getLifecycle().getCurrentState().isAtLeast(Lifecycle.State.STARTED)) {
+                    searching = query;
+                    adapter.search(query);
+                }
                 return true;
             }
         });