+++ app/src/main/java/eu/faircode/email/FragmentOptionsMisc.java
@@ -245,8 +245,9 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
+        Locale slocale = Resources.getSystem().getConfiguration().locale;
         for (String tag : getResources().getAssets().getLocales())
-            languages.add(new Pair<>(tag, Locale.forLanguageTag(tag).getDisplayName()));
+            languages.add(new Pair<>(tag, Locale.forLanguageTag(tag).getDisplayName(slocale)));
 
         Collections.sort(languages, new Comparator<Pair<String, String>>() {
             @Override
@@ -382,7 +383,7 @@ public class FragmentOptionsMisc extends FragmentBase implements SharedPreferenc
         ibHelp.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Helper.view(v.getContext(), Helper.getSupportUri(v.getContext()), false);
+                Helper.view(v.getContext(), Helper.getSupportUri(v.getContext(), "Options:misc"), false);
             }
         });